:truck: Move NewVersionModal to own directory
Moving to own directory so as to create new components for each step of
the flow.
